refactor(backend): rename credentials_auth.py to sso.py
Rename api/views/credentials_auth.py → api/views/sso.py and
tests/test_credentials_auth.py → tests/test_sso.py. Update all
import references.
